One file for each core command and a file for helper functions
